requirements.txt
Django>=5.0,<5.1
djangorestframework
djangorestframework-simplejwt
django-filter
celery
redis
psycopg2-binary
boto3
python-dotenv
Pillow
pdfplumber
PyMuPDF
requests
dj-database-url
pytest
pytest-django
factory-boy
pytest-cov
reportlab
